// src/enums.ts
export enum VariantType {
    NONE = 'NONE',
    SUCCESS = 'SUCCESS',
    WARNING = 'WARNING',
    DANGER = 'DANGER',
}

export enum SizeType {
    EXTRA_SMALL = 'extra_small',
    SMALL = 'small',
    MEDIUM = 'medium',
    LARGE = 'large'
}

export enum WidthType {
    FLUID = 'fluid',
    FIXED = 'fixed',
}